# apps/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Our Haven API
  version: 0.0.0
  description: Our Haven v1 API — Parent + Provider mobile, Provider web portal,
    admin dashboard. Source of truth per ADR-0004 (§§ 1–3, 8) and ADR-0010
    (Supabase + Fly.io + Vercel platform). US-region only.
  license:
    name: Proprietary
components:
  securitySchemes:
    supabaseAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (HS256). Verified locally with the
        project JWT secret on every request.
...
paths:
...
  /v1/admin/stripe-tax/preview-calculation:
    post:
      summary: Compute a Stripe Tax preview for a given purpose + US state + amount.
        Persists an audit row.
      tags:
        - admin
        - tax
      description: "Calls Stripe Tax `POST /v1/tax/calculations` for either
        `subscription` (state = subscriber's resident state) or `commission`
        (state = Provider's resident state). Persists the result to
        `stripe_tax_calculations`. Used during launch verification to sample tax
        outcomes across multiple states (OH-111 ACs #1 + #2) and at runtime as a
        preview before the Subscription/Commission flow actually charges. The
        purpose enum deliberately excludes Bookings — Our Haven does not collect
        sales tax on Bookings (CONTEXT.md § Sales tax model)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                purpose:
                  type: string
                  enum:
                    - subscription
                    - commission
                amountCents:
                  type: integer
                  exclusiveMinimum: 0
                  maximum: 9007199254740991
                state:
                  type: string
                  minLength: 2
                  maxLength: 2
                  pattern: ^[A-Z]{2}$
                postalCode:
                  type: string
                  minLength: 3
                  maxLength: 10
                city:
                  type: string
                  maxLength: 80
                reference:
                  type: string
                  minLength: 1
                  maxLength: 120
                subjectUid:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              required:
                - purpose
                - amountCents
                - state
                - reference
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  calculationId:
                    type: string
                  stripeCalculationId:
                    type: string
                  purpose:
                    type: string
                    enum:
                      - subscription
                      - commission
                  customerState:
                    type: string
                    minLength: 2
                    maxLength: 2
                    pattern: ^[A-Z]{2}$
                  customerPostalCode:
                    anyOf:
                      - type: string
                      - type: "null"
                  amountCents:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  taxAmountCents:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  amountTotalCents:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  taxBehavior:
                    type: string
                    enum:
                      - inclusive
                      - exclusive
                  taxCode:
                    type: string
                  taxBreakdown:
                    type: array
                    items:
                      type: object
                      properties:
                        amount:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        inclusive:
                          type: boolean
                        taxabilityReason:
                          anyOf:
                            - type: string
                            - type: "null"
                        taxableAmount:
                          anyOf:
                            - type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                            - type: "null"
                        state:
                          anyOf:
                            - type: string
                            - type: "null"
                        taxType:
                          anyOf:
                            - type: string
                            - type: "null"
                        percentageDecimal:
                          anyOf:
                            - type: string
                            - type: "null"
                      required:
                        - amount
                        - inclusive
                        - taxabilityReason
                        - taxableAmount
                        - state
                        - taxType
                        - percentageDecimal
                      additionalProperties: false
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - calculationId
                  - stripeCalculationId
                  - purpose
                  - customerState
                  - customerPostalCode
                  - amountCents
                  - taxAmountCents
                  - amountTotalCents
                  - taxBehavior
                  - taxCode
                  - taxBreakdown
                  - expiresAt
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "502":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
  /v1/admin/stripe-tax/calculations:
    get:
      summary: List recent Stripe Tax calculation audit rows.
      tags:
        - admin
        - tax
      description: Reads the `stripe_tax_calculations` audit table. Filterable by
        purpose, customer state, and subject uid. Returns up to 100 rows ordered
        by creation time (newest first).
      parameters:
        - schema:
            type: string
            enum:
              - subscription
              - commission
          in: query
          name: purpose
          required: false
        - schema:
            type: string
            minLength: 2
            maxLength: 2
            pattern: ^[A-Z]{2}$
          in: query
          name: state
          required: false
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: query
          name: subjectUid
          required: false
        - schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 100
          in: query
          name: limit
          required: false
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  calculations:
                    type: array
                    items:
                      type: object
                      properties:
                        calculationId:
                          type: string
                        stripeCalculationId:
                          type: string
                        purpose:
                          type: string
                          enum:
                            - subscription
                            - commission
                        customerState:
                          type: string
                          minLength: 2
                          maxLength: 2
                          pattern: ^[A-Z]{2}$
                        customerPostalCode:
                          anyOf:
                            - type: string
                            - type: "null"
                        amountCents:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        taxAmountCents:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        amountTotalCents:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        taxBehavior:
                          type: string
                          enum:
                            - inclusive
                            - exclusive
                        taxCode:
                          type: string
                        taxBreakdown:
                          type: array
                          items:
                            type: object
                            properties:
                              amount:
                                type: integer
                                minimum: -9007199254740991
                                maximum: 9007199254740991
                              inclusive:
                                type: boolean
                              taxabilityReason:
                                anyOf:
                                  - type: string
                                  - type: "null"
                              taxableAmount:
                                anyOf:
                                  - type: integer
                                    minimum: -9007199254740991
                                    maximum: 9007199254740991
                                  - type: "null"
                              state:
                                anyOf:
                                  - type: string
                                  - type: "null"
                              taxType:
                                anyOf:
                                  - type: string
                                  - type: "null"
                              percentageDecimal:
                                anyOf:
                                  - type: string
                                  - type: "null"
                            required:
                              - amount
                              - inclusive
                              - taxabilityReason
                              - taxableAmount
                              - state
                              - taxType
                              - percentageDecimal
                            additionalProperties: false
                        expiresAt:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                      required:
                        - calculationId
                        - stripeCalculationId
                        - purpose
                        - customerState
                        - customerPostalCode
                        - amountCents
                        - taxAmountCents
                        - amountTotalCents
                        - taxBehavior
                        - taxCode
                        - taxBreakdown
                        - expiresAt
                      additionalProperties: false
                required:
                  - calculations
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: https://api.ourhaven.example
    description: Production (Fly.io iad)
  - url: http://localhost:8080
    description: Local development
...